Refactor status and card components with improved structure
- Move ErrorCard and LoadingCard to dedicated status/cards directory
- Update import paths for status and card components
- Enhance SwipeFinishCard with ThemedText and PrimaryButton
- Reorganize status component exports
