+++ feature/links/presentation/components/display/cards/SwipeFinishCard.tsx
@@ -1,4 +1,7 @@
-import { Text, TouchableOpacity, View } from "react-native";
+import { View } from "react-native";
+
+import { PrimaryButton } from "@/components/button/PrimaryButton";
+import { ThemedText } from "@/components/text/ThemedText";
 
 interface SwipeFinishCardProps {
   onReload: () => void;
@@ -7,15 +10,23 @@ interface SwipeFinishCardProps {
 export function SwipeFinishCard({ onReload }: SwipeFinishCardProps) {
   return (
     <View className="flex-1 items-center justify-center">
-      <Text className="mb-4 text-xl font-medium text-gray-800">
-        You&apos;ve seen all cards
-      </Text>
-      <TouchableOpacity
-        onPress={onReload}
-        className="rounded-xl bg-accent px-8 py-4"
-      >
-        <Text className="text-base font-medium text-white">View Again</Text>
-      </TouchableOpacity>
+      <View className="flex flex-col items-center justify-center gap-2">
+        <ThemedText
+          text="You've seen all cards"
+          variant="body"
+          weight="normal"
+          color="default"
+        />
+        <PrimaryButton onPress={onReload}>
+          <ThemedText
+            text="View Again"
+            variant="body"
+            weight="normal"
+            color="white"
+            className="px-2"
+          />
+        </PrimaryButton>
+      </View>
     </View>
   );
 }
