+++ feature/links/presentation/components/display/cards/index.tsx
@@ -1,6 +1,6 @@
 export * from "./LinkInfoCard";
 export * from "./SwipeFinishCard";
 export * from "./FeaturedCard";
-export * from "./ErrorCard";
-export * from "./LoadingCard";
+export * from "../status/cards/ErrorCard";
+export * from "../status/cards/LoadingCard";
 export * from "./HorizontalCard";
