+++ feature/links/presentation/components/display/status/swipe/SwipeScreenStatus.tsx
